+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed not being possible to pass `undefined` as value to Bud, while readme allows that.
+
 ## `0.2.0` — 2019-11-30
 * **Breaking change**: now all effects run after all propagations, which means effects see all tip values in consistent state. Previously, effects run before propagating data further.